+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - accessibility page
 - add link to school details in GIAS from case view
 - add ability to re-categorise a case
+- enable sending of emails via Notify
 
 ### Specify Unreleased